# Optimization: coalesce payments before inserting