On mainnet there are a substantial number of addresses that receive payments from multiple SNs at each block.  Previously we were performing separate INSERT statements for each of these payments.

This commit optimizes it to accumulate all of the amounts to each address and then pays out the accumulated amounts, i.e. with one query per paid address instead of one per address per SN.

This makes syncing and subsystem rescanning a little more than twice as fast (once into the batching part of the chain, i.e. after HF19).  I go from syncing around 75-90 blocks/second to around 175-200 blocks per second, and rescanning about 2.2x faster (with a 2/3 reduction in the amount of time spent in the batch part of rescanning specifically):

Before:

    [+0.143s] [global:info|cryptonote_core/blockchain.cpp:374] Loading blocks into oxen subsystems, scanning blockchain from height: 1080001 to: 1100111 (snl: 1080001, ons: 1100111, sqlite: 1086858)
    [+0.143s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1080001 (0.000s) (snl: 0.000s, ons: 0.000s, batch: 0.000s)
    [+10.787s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1087401 (10.488s) (snl: 5.841s, ons: 0.000s, batch: 4.512s)
    [+21.468s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1088601 (10.674s) (snl: 0.925s, ons: 0.000s, batch: 9.730s)
    [+32.111s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1089801 (10.639s) (snl: 0.923s, ons: 0.000s, batch: 9.692s)
    [+42.289s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1090901 (10.012s) (snl: 0.856s, ons: 0.000s, batch: 9.140s)
    [+53.159s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1092101 (10.861s) (snl: 0.933s, ons: 0.000s, batch: 9.909s)
    [+1m03.207s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1093201 (10.044s) (snl: 0.876s, ons: 0.000s, batch: 9.148s)
    [+1m13.313s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1094301 (10.102s) (snl: 0.847s, ons: 0.000s, batch: 9.206s)
    [+1m23.448s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1095401 (10.129s) (snl: 0.877s, ons: 0.000s, batch: 9.208s)
    [+1m33.580s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1096501 (10.128s) (snl: 0.866s, ons: 0.000s, batch: 9.228s)
    [+1m43.672s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1097601 (10.086s) (snl: 0.904s, ons: 0.000s, batch: 9.166s)
    [+1m54.294s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1098801 (10.618s) (snl: 0.929s, ons: 0.000s, batch: 9.676s)
    [+2m04.828s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1100001 (10.529s) (snl: 0.913s, ons: 0.000s, batch: 9.606s)
    [+2m05.801s] [global:info|cryptonote_core/blockchain.cpp:497] Done recalculating oxen subsystems in 125.66s (15.69s snl; 0.00s ons; 108.22s batch)

After:

    [+0.141s] [global:info|cryptonote_core/blockchain.cpp:374] Loading blocks into oxen subsystems, scanning blockchain from height: 1080001 to: 1100111 (snl: 1080001, ons: 1100111, sqlite: 1086858)
    [+0.141s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1080001 (0.000s) (snl: 0.000s, ons: 0.000s, batch: 0.000s)
    [+10.586s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1088101 (10.297s) (snl: 6.404s, ons: 0.000s, batch: 3.745s)
    [+20.974s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1090801 (10.224s) (snl: 2.059s, ons: 0.000s, batch: 8.122s)
    [+31.329s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1093501 (10.348s) (snl: 2.096s, ons: 0.000s, batch: 8.206s)
    [+41.653s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1096101 (10.320s) (snl: 2.052s, ons: 0.000s, batch: 8.163s)
    [+51.854s] [global:info|cryptonote_core/blockchain.cpp:398] ... scanning height 1098701 (10.197s) (snl: 2.086s, ons: 0.000s, batch: 8.071s)
    [+57.376s] [global:info|cryptonote_core/blockchain.cpp:497] Done recalculating oxen subsystems in 57.23s (14.70s snl; 0.00s ons; 36.31s batch)

(These results were on my workstation with a reasonably fast, idle nvme drive; on slower drives I'd expect the relative speedup to be larger still).